add build constraints
